organize configuration.
